% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Connect to the Indicare Salute Lab database}
\usage{
db_connect(uid = Sys.getenv("UID"), pwd = Sys.getenv("PWD"))
}
\arguments{
\item{uid}{Character. Database username. Defaults to the \code{UID} environment variable.}

\item{pwd}{Character. Database password. Defaults to the \code{PWD} environment variable.}
}
\value{
A \code{DBIConnection} object.
}
\description{
Opens a connection to the \code{IndicareSaluteLab} Azure SQL database via ODBC.
Credentials are read from environment variables by default and can be set
in \code{.Renviron}.
}
\examples{
\dontrun{
con <- db_connect()
DBI::dbListTables(con)
DBI::dbDisconnect(con)
}
}
