% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restraintsfalls-package.R
\docType{package}
\name{restraintsfalls-package}
\alias{restraintsfalls}
\alias{restraintsfalls-package}
\title{restraintsfalls: Physical Restraint Use and Falls in Nursing Homes}
\description{
Research compendium for the analysis of the association between physical restraint use and falls in nursing homes in the province of Trento, Italy. Uses longitudinal quality monitoring data from the Indicare Salute Lab system (2016-2025).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://upipa.github.io/restraints-falls-nh/}
  \item Report bugs at \url{https://github.com/upipa/restraints-falls-nh/issues}
}

}
\author{
\strong{Maintainer}: Ivan Marino \email{ivan.marino@upipa.tn.it}

Authors:
\itemize{
  \item Ivan Marino \email{ivan.marino@upipa.tn.it}
}

}
\keyword{internal}
